+++ src/main/java/me/kall/doespotatotick/common/config/PotatoConfig.java
@@ -44,6 +44,7 @@ public class PotatoConfig {
     public static final ForgeConfigSpec.ConfigValue<List<? extends String>> DIMENSION_WHITELIST;
     public static final ForgeConfigSpec.BooleanValue STOP_RENDERING_SKIPPED_ENTITIES;
     public static final ForgeConfigSpec.IntValue ENTITY_RENDERABLE_REFRESH_INTERVAL;
+    public static final ForgeConfigSpec.BooleanValue USE_AVG_FPS_AS_REFRESH_INTERVAL;
 
     static {
         List<? extends String> itemList = Lists.newArrayList("minecraft:cobblestone");
@@ -90,6 +91,7 @@ public class PotatoConfig {
         builder.push("Client");
         STOP_RENDERING_SKIPPED_ENTITIES = builder.comment("If the tick of an entity is skipped by this mod, stop its client rendering so that players won't get it stuck in their worlds.").define("StopRenderingSkippedEntities", true);
         ENTITY_RENDERABLE_REFRESH_INTERVAL = builder.comment("How often the mod checks if an entity should be ticked and rendered, counted in rendering calls instead of game ticks. Higher values mean it checks less often, which might save a tiny bit of performance (honestly, not super noticeable lol), but updates will feel a little slower.").defineInRange("EntityRenderableRefreshInterval", 60,0,Integer.MAX_VALUE);
+        USE_AVG_FPS_AS_REFRESH_INTERVAL = builder.comment("Use client's average fps as EntityRenderableRefreshInterval.").define("UseAvgFpsAsRefreshInterval", true);
         builder.pop();
         COMMON_CONFIG = builder.build();
     }