+++ src/main/java/me/kall/doespotatotick/common/mixin/EntityMixin.java
@@ -3,9 +3,9 @@ package me.kall.doespotatotick.common.mixin;
 import me.kall.doespotatotick.DoesPotatoTick;
 import me.kall.doespotatotick.common.api.Tickable;
 import me.kall.doespotatotick.common.config.PotatoConfig;
+import me.kall.doespotatotick.common.fps.FpsViewer;
 import net.minecraft.world.entity.Entity;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -13,24 +13,22 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Entity.class)
 public abstract class EntityMixin implements Tickable {
-    @Shadow public abstract net.minecraft.world.level.Level level();
-    @Unique private int client$tickCount;
-    @Unique private volatile boolean client$isTickable;
+    @Unique private int client$renderCallCount;
+    @Unique private volatile boolean client$isRenderable;
 
     @Override
-    public boolean doesPotatoTick$isTickable() {
-        return this.client$isTickable;
+    public boolean doesPotatoTick$isRenderable() {
+        return this.client$isRenderable;
     }
 
     @Inject(method = "shouldRender", at = @At("RETURN"))
-    private void onClientTick(CallbackInfoReturnable<Boolean> cir) {
-        if (this.level().isClientSide()) {
-            this.client$tickCount++;
-            int interval = PotatoConfig.ENTITY_RENDERABLE_REFRESH_INTERVAL.get();
-            if (this.client$tickCount >= interval) {
-                this.client$tickCount = 0;
-
-                this.client$isTickable = DoesPotatoTick.isTickable((Entity) (Object) this);
+    private void onRenderCall(CallbackInfoReturnable<Boolean> cir) {
+        if (PotatoConfig.STOP_RENDERING_SKIPPED_ENTITIES.get()) {
+            this.client$renderCallCount++;
+            int interval = FpsViewer.getAvgFps();
+            if (this.client$renderCallCount >= interval) {
+                this.client$renderCallCount = 0;
+                this.client$isRenderable = DoesPotatoTick.isTickable((Entity) (Object) this);
             }
         }
     }