+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-# 5.3.1
-- add config comment for StopRenderingSkippedEntities option
-- Add EntityRenderableRefreshInterval config option (with comment)
+# 5.3.2
+- add UseAvgFpsAsRefreshInterval config option (so that basically users don't need to tweak the client config options as it would have nothing to do with their game experience lol)